+++ controller/media_files_controller.js
@@ -317,14 +317,23 @@ MediaFiles.setAction(function uploadsingle(conduit) {
 /**
  * Get the file data
  *
- * @author   Jelle De Loecker   <jelle@develry.be>
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.5.1
- * @version  0.5.1
+ * @version  0.6.4
  *
- * @param    {Conduit}               conduit
- * @param    {Document.MediaFile}    media_file
+ * @param    {Conduit}     conduit
+ * @param    {ObjectId}    media_file_id
  */
-MediaFiles.setAction(function data(conduit, media_file) {
+MediaFiles.setAction(async function data(conduit, prefix, media_file_id) {
+
+	conduit.prefix = prefix;
+
+	let media_file = await this.model.findById(media_file_id);
+
+	if (!media_file) {
+		return conduit.notFound();
+	}
+
 	conduit.setHeader('cache-control', 'public, max-age=3600, must-revalidate');
 
 	conduit.end({